requirements.txt
python-gnupg==0.3.5
redis==2.8.0
